+++ frontend/src/components/staticpages/DockerImagePage.tsx
