+++ frontend/src/components/staticpages/LandingPage.tsx
